+++ tools/src/main/kotlin/com/dreampany/tools/data/source/repository/NoteRepository.kt
@@ -96,11 +96,11 @@ class NoteRepository
         //return concatFirstRx(true, room)
     }
 
-    override fun getItems(limit: Int): List<Note>? {
+    override fun getItems(limit: Long): List<Note>? {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 
-    override fun getItemsRx(limit: Int): Maybe<List<Note>> {
+    override fun getItemsRx(limit: Long): Maybe<List<Note>> {
         TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
     }
 }
